pyproject.toml
[project]
name = "strands-agentcore-tf"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "strands-agents>=1.15.0",
    "uvicorn[standard]>=0.38.0",
]
